feat(P15-F03): extract ExpenseForm into its own component
Moves the inline ExpenseForm function, its option lists, and field
mapping out of MonthlyPage.tsx into src/components/ExpenseForm.tsx,
following F02's grid-extraction precedent. No behavior change.
